requirements.txt
Pillow==9.1.1
PyPDF2==3.0.1
pytesseract
openai==0.27.8
pdf2image==1.16.3
python-dotenv==1.0.0